# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ArchiveChatRequest:
    properties:
      archive:
        example: true
        type: boolean
      phone:
        example: "5511999999999"
        type: string
    required:
    - phone
    type: object
...
  dto.ChatActionResponse:
    properties:
      status:
        example: archived
        type: string
    type: object
...
  dto.CommunityResponse:
    properties:
      groups: {}
    type: object
...
  dto.DeleteMessageRequest:
    properties:
      forMe:
        example: false
        type: boolean
      messageId:
        example: ABCD1234
        type: string
      phone:
        example: "5511999999999"
        type: string
    required:
    - messageId
    - phone
    type: object
  dto.DisappearingRequest:
    properties:
      phone:
        example: "5511999999999"
        type: string
      timer:
        description: 24h, 7d, 90d, or off
        example: 24h
        type: string
    required:
    - phone
    - timer
    type: object
  dto.EditMessageRequest:
    properties:
      messageId:
        example: ABCD1234
        type: string
      newText:
        example: Edited message
        type: string
      phone:
        example: "5511999999999"
        type: string
    required:
    - messageId
    - newText
    - phone
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: session not found
        type: string
    type: object
...
  dto.LinkGroupRequest:
    properties:
      childGroupId:
        example: 987654321@g.us
        type: string
      parentGroupId:
        example: 123456789@g.us
        type: string
    required:
    - childGroupId
    - parentGroupId
    type: object
...
  dto.MessageResponse:
    properties:
      id:
        example: ABCD1234
        type: string
      message:
        example: operation completed
        type: string
      status:
        example: connected
        type: string
    type: object
...
  dto.RejectCallRequest:
    properties:
      callFrom:
        example: "5511999999999"
        type: string
      callId:
        example: CALL123
        type: string
    required:
    - callFrom
    - callId
    type: object
  dto.RequestUnavailableMessageRequest:
    properties:
      senderJid:
        example: 5511999999999@s.whatsapp.net
        type: string
    type: object
...
  dto.SendResponse:
    properties:
      messageId:
        example: ABCD1234
        type: string
      timestamp:
        example: 1699999999
        type: integer
    type: object
...
info:
  contact: {}
  description: WhatsApp API with Chatwoot Integration
  title: ZPWoot WhatsApp API
  version: 0.1.0
paths:
...
  /sessions/{sessionId}/calls/reject:
    post:
      consumes:
      - application/json
      description: Reject an incoming WhatsApp call
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Call data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RejectCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Reject incoming call
      tags:
      - call
  /sessions/{sessionId}/chats/{chatId}/archive:
    patch:
      consumes:
      - application/json
      description: Archive or unarchive a WhatsApp chat
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Archive data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArchiveChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ChatActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Archive or unarchive a chat
      tags:
      - chat
  /sessions/{sessionId}/chats/{chatId}/messages/{messageId}:
    delete:
      consumes:
      - application/json
      description: Delete a message from a chat
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Delete data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a message
      tags:
      - chat
    patch:
      consumes:
      - application/json
      description: Edit a previously sent message
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Edit data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.EditMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Edit a message
      tags:
      - chat
  /sessions/{sessionId}/chats/{chatId}/messages/{messageId}/request:
    post:
      consumes:
      - application/json
      description: Request WhatsApp to resend an unavailable message from phone storage
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID
        in: path
        name: chatId
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      - description: Optional sender JID for group messages
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.RequestUnavailableMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Request unavailable message
      tags:
      - chat
...
  /sessions/{sessionId}/chats/{chatId}/settings/disappearing:
    patch:
      consumes:
      - application/json
      description: Set the disappearing messages timer for a chat
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Timer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DisappearingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set disappearing messages timer
      tags:
      - chat
...
  /sessions/{sessionId}/communities/{communityId}/groups:
    get:
      description: Get list of subgroups in a community
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Community ID
        in: path
        name: communityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommunityResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get community subgroups
      tags:
      - community
    post:
      consumes:
      - application/json
      description: Link a group to a community as a subgroup
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Link data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LinkGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Link group to community
      tags:
      - community
  /sessions/{sessionId}/communities/{communityId}/groups/{groupId}:
    delete:
      consumes:
      - application/json
      description: Unlink a group from a community
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Unlink data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LinkGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unlink group from community
      tags:
      - community
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"
...